requirements.txt
opencv-python ==  4.5.3.56
ffmpeg-python == 0.2
batch_face == 1.0.0